+++ app/router/signup.test.js
@@ -14,6 +14,7 @@ jest.mock('axios');
 const store = createStore(rootReducer, applyMiddleware(thunk));
 
 describe('routes/signup', () => {
+  const mockValidateRoles = jest.fn();
   const roles = {
     pending: false,
     success: true,
@@ -23,7 +24,7 @@ describe('routes/signup', () => {
     return mount(
       <Provider store={store}>
         <MemoryRouter initialEntries={[route]}>
-          <App t={t} roles={roles} />
+          <App t={t} validateRoles={mockValidateRoles} roles={roles} />
         </MemoryRouter>
       </Provider>
     );
