+++ package.json
@@ -18,10 +18,11 @@
   },
   "dependencies": {
     "@babel/register": "^7.7.0",
+    "@helix-design-system/helix-react": "^1.3.0",
     "@webcomponents/webcomponentsjs": "^2.4.0",
     "axios": "^0.19.0",
     "fs": "^0.0.1-security",
-    "helix-ui": "^0.20.0",
+    "helix-ui": "^0.24.0",
     "i18next": "^19.0.0",
     "i18next-browser-languagedetector": "^4.0.1",
     "i18next-icu": "^1.1.2",
